Added user authentication in all APIs news/event/notice

+++ event/views.py
@@ -5,6 +5,9 @@ from rest_framework.decorators import list_route
 from serializers import EventReadSerializer, EventLikeSerializer, EventViewSerializer
 from rest_framework.response import Response
 from rest_framework.status import HTTP_400_BAD_REQUEST
+from rest_framework.permissions import IsAuthenticated
+from iitbapp.permissions import IsCorrectUserId
+from authentication.tokenauth import TokenAuthentication
 
 
 class EventPagination(LimitOffsetPagination):
@@ -16,6 +19,8 @@ class EventViewSet(viewsets.ReadOnlyModelViewSet):
     queryset = Event.objects.all().order_by('-id')
     pagination_class = EventPagination
     serializer_class = EventReadSerializer
+    authentication_classes = [TokenAuthentication]
+    permission_classes = [IsAuthenticated, IsCorrectUserId]
 
     @list_route(methods=['POST'])
     def like(self, request):
@@ -23,6 +28,7 @@ class EventViewSet(viewsets.ReadOnlyModelViewSet):
         if serializer.is_valid():
             event = serializer.data['event']
             user = serializer.data['user']
+            self.check_object_permissions(request, user)
             event_like, created = EventLike.objects.get_or_create(event__id=event, user__id=user,
                                                                   defaults={'event_id': event, 'user_id': user})
             return Response(EventLikeSerializer(event_like).data)
@@ -35,6 +41,7 @@ class EventViewSet(viewsets.ReadOnlyModelViewSet):
         if serializer.is_valid():
             event = serializer.data['event']
             user = serializer.data['user']
+            self.check_object_permissions(request, user)
             EventLike.objects.all().filter(event=event).filter(user=user).delete()
             likes = EventLike.objects.all().filter(news=event).count()
             return Response({'user': user, 'event': event, 'id': -1, 'likes': likes})
@@ -47,6 +54,7 @@ class EventViewSet(viewsets.ReadOnlyModelViewSet):
         if serializer.is_valid():
             event = serializer.data['event']
             user = serializer.data['user']
+            self.check_object_permissions(request, user)
             event_view, created = EventViews.objects.get_or_create(event__id=event, user_id=user,
                                                                    defaults={'event_id': event, 'user_id': user})
             return Response(EventViewSerializer(event_view).data)
